added: table CustomerType, Customer, ContentClass, Context, CustomerContext
