+++ backend/repetitor_backend/piccolo_app.py
@@ -2,9 +2,11 @@ import os
 
 from piccolo.conf.apps import AppConfig
 from .tables import (
-    Customer,
     CustomerType,
-    ContentClass, Content
+    Customer,
+    ContentClass,
+    Context,
+    CustomerContext,
 )
 
 
@@ -12,9 +14,12 @@ CURRENT_DIRECTORY = os.path.dirname(os.path.abspath(__file__))
 
 
 APP_CONFIG = AppConfig(
-    app_name="repetitor_backend",
-    migrations_folder_path=os.path.join(CURRENT_DIRECTORY, "piccolo_migrations"),
-    table_classes=[Customer, CustomerType, ContentClass, Content],
+    app_name='repetitor_backend',
+    migrations_folder_path=os.path.join(
+        CURRENT_DIRECTORY,
+        'piccolo_migrations'
+    ),
+    table_classes=[CustomerType, Customer, ContentClass, Context, CustomerContext],
     migration_dependencies=[],
-    commands=[],
+    commands=[]
 )
