--- backend/repetitor_backend/piccolo_migrations/repetitor_backend_2023_05_17t17_16_01_890314.py
@@ -1,17 +1,0 @@
-from piccolo.apps.migrations.auto.migration_manager import MigrationManager
-
-
-ID = "2023-05-17T17:16:01:890314"
-VERSION = "0.111.0"
-DESCRIPTION = ""
-
-
-async def forwards():
-    manager = MigrationManager(migration_id=ID, app_name="", description=DESCRIPTION)
-
-    def run():
-        print(f"running {ID}")
-
-    manager.add_raw(run)
-
-    return manager
