--- backend/repetitor_backend/piccolo_migrations/repetitor_backend_2023_05_17t17_17_11_850003.py
@@ -1,292 +1,0 @@
-from piccolo.apps.migrations.auto.migration_manager import MigrationManager
-from piccolo.columns.base import OnDelete
-from piccolo.columns.base import OnUpdate
-from piccolo.columns.column_types import BigInt
-from piccolo.columns.column_types import Boolean
-from piccolo.columns.column_types import ForeignKey
-from piccolo.columns.column_types import UUID
-from piccolo.columns.column_types import Varchar
-from piccolo.columns.defaults.uuid import UUID4
-from piccolo.columns.indexes import IndexMethod
-from piccolo.table import Table
-
-
-class CustomerClass(Table, tablename="customer_class"):
-    id = UUID(
-        default=UUID4(),
-        null=False,
-        primary_key=True,
-        unique=False,
-        index=False,
-        index_method=IndexMethod.btree,
-        choices=None,
-        db_column_name=None,
-        secret=False,
-    )
-
-
-ID = "2023-05-17T17:17:11:850003"
-VERSION = "0.111.0"
-DESCRIPTION = ""
-
-
-async def forwards():
-    manager = MigrationManager(
-        migration_id=ID, app_name="repetitor_backend", description=DESCRIPTION
-    )
-
-    manager.add_table("Customer", tablename="customer")
-
-    manager.add_table("CustomerClass", tablename="customer_class")
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="id",
-        db_column_name="id",
-        column_class_name="UUID",
-        column_class=UUID,
-        params={
-            "default": UUID4(),
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="customer_class",
-        db_column_name="customer_class",
-        column_class_name="ForeignKey",
-        column_class=ForeignKey,
-        params={
-            "references": CustomerClass,
-            "on_delete": OnDelete.cascade,
-            "on_update": OnUpdate.cascade,
-            "target_column": None,
-            "null": True,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="telegram_user_id",
-        db_column_name="telegram_user_id",
-        column_class_name="BigInt",
-        column_class=BigInt,
-        params={
-            "default": 0,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="telegram_language",
-        db_column_name="telegram_language",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "lenght": 3,
-            "length": 255,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="user_name",
-        db_column_name="user_name",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 50,
-            "default": "",
-            "null": True,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="first_name",
-        db_column_name="first_name",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 50,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="last_name",
-        db_column_name="last_name",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 50,
-            "default": "",
-            "null": True,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Customer",
-        tablename="customer",
-        column_name="is_active",
-        db_column_name="is_active",
-        column_class_name="Boolean",
-        column_class=Boolean,
-        params={
-            "default": True,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="CustomerClass",
-        tablename="customer_class",
-        column_name="id",
-        db_column_name="id",
-        column_class_name="UUID",
-        column_class=UUID,
-        params={
-            "default": UUID4(),
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="CustomerClass",
-        tablename="customer_class",
-        column_name="name",
-        db_column_name="name",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 50,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="CustomerClass",
-        tablename="customer_class",
-        column_name="describe",
-        db_column_name="describe",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 200,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="CustomerClass",
-        tablename="customer_class",
-        column_name="is_active",
-        db_column_name="is_active",
-        column_class_name="Boolean",
-        column_class=Boolean,
-        params={
-            "default": True,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    return manager
