--- backend/repetitor_backend/piccolo_migrations/repetitor_backend_2023_05_26t16_14_30_510450.py
@@ -1,103 +1,0 @@
-from piccolo.apps.migrations.auto.migration_manager import MigrationManager
-from piccolo.columns.column_types import Boolean
-from piccolo.columns.column_types import Text
-from piccolo.columns.column_types import UUID
-from piccolo.columns.column_types import Varchar
-from piccolo.columns.defaults.uuid import UUID4
-from piccolo.columns.indexes import IndexMethod
-
-
-ID = "2023-05-26T16:14:30:510450"
-VERSION = "0.111.0"
-DESCRIPTION = ""
-
-
-async def forwards():
-    manager = MigrationManager(
-        migration_id=ID, app_name="repetitor_backend", description=DESCRIPTION
-    )
-
-    manager.add_table("ContentClass", tablename="content_class")
-
-    manager.add_column(
-        table_class_name="ContentClass",
-        tablename="content_class",
-        column_name="id",
-        db_column_name="id",
-        column_class_name="UUID",
-        column_class=UUID,
-        params={
-            "default": UUID4(),
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="ContentClass",
-        tablename="content_class",
-        column_name="description",
-        db_column_name="description",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="ContentClass",
-        tablename="content_class",
-        column_name="name",
-        db_column_name="name",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 20,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="ContentClass",
-        tablename="content_class",
-        column_name="is_active",
-        db_column_name="is_active",
-        column_class_name="Boolean",
-        column_class=Boolean,
-        params={
-            "default": True,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    return manager
